pyproject.toml
[tool.poetry]
name = "rooms-test"
version = "0.1.0"
description = ""
authors = ["Martin Dluhoš <dluhos.martin@gmail.com>"]
readme = "README.md"
packages = [{include = "rooms_test"}]

[tool.poetry.dependencies]
python = "^3.9"
google-api-python-client = "^2.78.0"
google-auth-oauthlib = "^1.0.0"
colorama = "^0.4.6"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
